\name{EPICv2manifest-package}
\alias{EPICv2manifest-package}
\docType{package}
\title{
Extended EPICv2 manifest data package 
}
\description{
\code{EPICv2manifest} is an annotation package providing a data.frame object containing the complete 49 column Infinium Illumina EPIC v2.0 probe manifest (``IlmnID'' constitutes the rownames), plus 31 additional columns derived from Peters et al. (2024). 
}
\author{
Tim Peters, Immunogenomics Laboratory, Garvan Institute of Medical Research

Maintainer: Tim Peters<t.peters@garvan.org.au>
}

\examples{
data(EPICv2manifest)
}

\keyword{datasets}
