\name{EPICv2manifest}
\docType{data}
\alias{EPICv2manifest}
\title{Extended Illumina EPICv2 manifest}
\description{
  A data.frame containing an extended probe manifest for the Illumina Infinium Methylation v2.0 Kit. Contains the complete manifest from the Illumina-provided EPIC-8v2-0_EA.csv, plus additional probewise information described in Peters et al. (2024).
}
\usage{data(EPICv2manifest)}
\value{A data.frame containing 80 columns, with Illumina IDs as rownames}
\format{data.frame}
\source{"https://bioconductor.org/packages/release/data/annotation/html/EPICv2manifest.html"}
\keyword{datasets}