\name{EPICv2manifest-package}
\alias{EPICv2manifest-package}
\docType{package}
\title{
Extended EPICv2 manifest data package 
}
\description{
\code{EPICv2manifest} is an annotation package providing a data.frame object containing the complete 49 column Infinium Illumina EPIC v2.0 probe manifest (``IlmnID'' constitutes the rownames), plus 31 additional columns derived from Peters et al. (2024). 
}
\author{
Tim Peters, Immunogenomics Laboratory, Garvan Institute of Medical Research

Maintainer: Tim Peters<t.peters@garvan.org.au>
}

\references{
Peters, T.J., Meyer, B., Ryan, L., Achinger-Kawecka, J., Song, J., Campbell, E.M., Qu, W., Nair, S., Loi-Luu, P., Stricker, P., Lim, E., Stirzaker, C., Clark, S.J. and Pidsley, R. (2024). Characterisation and reproducibility of the HumanMethylationEPIC v2.0 BeadChip for DNA methylation profiling. \emph{BMC Genomics} \bold{25}, 251. doi:10.1186/s12864-024-10027-5.
}

\examples{
data(EPICv2manifest)
}

\keyword{datasets}
