\name{EPICv2manifest}
\docType{data}
\alias{EPICv2manifest}
\title{Extended Illumina EPICv2 manifest}
\description{
  A data.frame containing an extended probe manifest for the Illumina Infinium Methylation v2.0 Kit. Contains the complete manifest from the Illumina-provided EPIC-8v2-0_EA.csv, plus additional probewise information described in Peters et al. (2024).
}
\usage{data(EPICv2manifest)}
\value{A data.frame containing 80 columns, with Illumina IDs as rownames}
\format{data.frame}
\references{
Peters, T.J., Meyer, B., Ryan, L., Achinger-Kawecka, J., Song, J., Campbell, E.M., Qu, W., Nair, S., Loi-Luu, P., Stricker, P., Lim, E., Stirzaker, C., Clark, S.J. and Pidsley, R. (2024). Characterisation and reproducibility of the HumanMethylationEPIC v2.0 BeadChip for DNA methylation profiling. BMC Genomics \bold{25}, 251. doi:10.1186/s12864-024-10027-5.
}
\source{\url{https://static-content.springer.com/esm/art\%3A10.1186\%2Fs12864-024-10027-5/MediaObjects/12864_2024_10027_MOESM4_ESM.csv}}
\keyword{datasets}
